refactor to break BaseResourceObject up
move resource objects to the models folder
pull out the static methods of BaseResourceObject into a resource manager module
move S3ResourceObject and IotResourceObject into their own files
add additional tests to cover error messages,  resource-manager has 98% coverage now
